Some improvements related to Grafana's annotations
1. Add documentation.
2. An event handles now a 'title', a 'text' and 'tags':
    * title defines the event type (playbook, play, task);
    * text defines the event name;
    * tags contain further information about the event:
        * event type;
        * user;
        * node;
        * Kolla ref, distro and install type;
        * Kolla tag;
        * task timer;
        * OpenStack release;
        * region name.
3. Because recording every task saved too many useless information, tasks are
no longer recorded when they begin. In order to record only the tasks that
imply a modification (marked as 'changed' in Ansible), this commit checks the
status of tasks before adding them to InfluxDB.
4. InfluxDB connexion is tested, if not working, the plugin is disabled
and a warning message appears.
5. Some parameters are provided to kolla-ansible such as `influx_vip`
which is required to record events during the `os_install` phase.